switch from Handle to Actor
because I have to interleave IO actions and state changes, I cannot make
my "atomically" blocks as big as I want them to be. as a result, I worry
that concurrent events will modify the state in unexpected ways.

by switching to an actor who only answers one message at a time, the
state is safely hidden inside the actor and thus cannot be changed in
unexpected ways.